+++ src/front/js/pages/NotFoundPage.jsx
@@ -1,6 +1,5 @@
 'use strict';
 import React, { useEffect } from 'react';
-import Metas from '../components/Metas';
 
 const NotFoundPage = () => {