+++ src/front/js/pages/SinglePostPage.jsx
@@ -1,7 +1,7 @@
 'use strict';
 import React, { useEffect } from 'react';
 
-const HomePage = ({ id, title, path, acf }) => {
+const SinglePostPage = ({ id, title, path, acf }) => {
     
     
 
@@ -17,16 +17,16 @@ const HomePage = ({ id, title, path, acf }) => {
 	
 	return(
 		<>
-			<section id="h__intro">
+			<section id="sp__intro">
 				<div className="container">
 					
 				</div>
 			</section>
+			<section style={{ background: '#0000ff', height: '100lvh' }}></section>
 			<section style={{ background: '#00ff00', height: '100lvh' }}></section>
-			<section style={{ background: '#000000', height: '100lvh' }}></section>
 		</>
 	);
 	
 }
 
-export default HomePage;
+export default SinglePostPage;