+++ src/front/js/pages/WaitingPage.jsx
@@ -1,6 +1,5 @@
 'use strict';
 import React, { useEffect } from 'react';
-import Metas from '../components/Metas';
 
 const WaitingPage = () => {
     
@@ -17,9 +16,7 @@ const WaitingPage = () => {
 	
 	return(
 		<>
-			<Metas
-				title={window.defaultMetas.siteName}
-			/>
+			
 		</>
 	);